+++ CHANGELOG.md
@@ -10,6 +10,11 @@ Full Keycloak upstream jira issue can be shown if filtered by Fix version.
 
 Our Keycloak version is working well with PostgreSQL database. For using other SQL databases, text field in database need to be evaluated.
 
+## [Unreleased]
+
+### Changed
+- SAML Federation process reexecute when no all IDPs are added
+
 ## [22.0.11-1.9] - 2024-10-10
 
 ### Fixed